Moved all the tests to separate folders
