+++ src/html/CSS/Index.js
@@ -1,5 +1,5 @@
 // @flow
-import Result, { TYPE_OK, TYPE_INCORRECT, TYPE_ERROR } from '../pentest/Result';
+import Result, { TYPE_OK, TYPE_INCORRECT, TYPE_ERROR } from '../../pentest/Result';
 import csso from 'csso';
 
 export default class CSS {
