+++ src/html/Image/Index.js
@@ -1,8 +1,7 @@
 // @flow
-import Result, { TYPE_OK, TYPE_INCORRECT, TYPE_ERROR } from '../pentest/Result';
-import UglifyJS from 'uglify-js';
+import Result, { TYPE_OK, TYPE_INCORRECT, TYPE_ERROR } from '../../pentest/Result';
 
-export default class JavaScript {
+export default class Image {
   request: Object;
   parseHtml: Function;
   getResources: Function;
@@ -22,10 +21,13 @@ export default class JavaScript {
     this.logger = logger;
   }
 
-  filterJavaScriptFiles(resources) {
-    const regx = new RegExp('.*\\/([a-z0-9.])+(\\.js)((\\?)?(.*))', 'i');
+  filterImageFiles(resources) {
+    const regxJpg = new RegExp('.*\\/([a-z0-9.-])+(\\.jpg)((\\?)?(.*))', 'i');
+    const regxJpeg = new RegExp('.*\\/([a-z0-9.-])+(\\.jpeg)((\\?)?(.*))', 'i');
+    const regxPng = new RegExp('.*\\/([a-z0-9.-])+(\\.png)((\\?)?(.*))', 'i');
+    const regxSvg = new RegExp('.*\\/([a-z0-9.-])+(\\.svg)((\\?)?(.*))', 'i');
     return resources.filter((resource) => {
-      return regx.test(resource);
+      return regxJpg.test(resource) || regxJpeg.test(resource) || regxPng.test(resource) || regxSvg.test(resource);
     });
   }
 
@@ -44,23 +46,7 @@ export default class JavaScript {
     return result.response.headers.hasOwnProperty('x-content-type-options');
   }
 
-  hasConsoleLogs(result: Object): boolean {
-    return result.body.indexOf('console.log') !== -1 && result.body.indexOf('console.error') !== -1;
-  }
-
-  isMinified(result: Object): boolean {
-    const r = UglifyJS.minify(result.body, { mangle: false, output: { comments: true } });
-
-    return result.body.length/* - (result.body.length * 0.05)*/ <= r.code.length;
-  }
-
-  isUglified(result: Object): boolean {
-    const r = UglifyJS.minify(result.body);
-  
-    return result.body.length/* - (result.body.length * 0.02)*/ <= r.code.length;
-  }
-
-  async checkJavaScriptFile(url: string): Object {
+  async checkImageFile(url: string): Object {
     const result = await this.request.get(url);
     let tests = [];
 
@@ -72,9 +58,6 @@ export default class JavaScript {
     tests = [...tests, new Result('available', TYPE_OK)];
     tests = [...tests, new Result('cache', this.isCached(result) ? TYPE_OK : TYPE_INCORRECT)];
     tests = [...tests, new Result('x-content-type-options', this.hasXContentTypeOptionsHeader(result) ? TYPE_OK : TYPE_INCORRECT)];
-    tests = [...tests, new Result('console-logs', !this.hasConsoleLogs(result) ? TYPE_OK : TYPE_INCORRECT)];
-    tests = [...tests, new Result('minify', this.isMinified(result) ? TYPE_OK : TYPE_INCORRECT)];
-    tests = [...tests, new Result('uglify', this.isUglified(result) ? TYPE_OK : TYPE_INCORRECT)];
 
     return new Result(url, tests.filter(test => test.resultType !== TYPE_OK).length === 0 ? TYPE_OK : TYPE_INCORRECT, tests);
   }
@@ -85,14 +68,13 @@ export default class JavaScript {
    * @returns {Object}
    */
   async execute(url: string): Object {
-    this.logger.verbose('Executing JavaScript test...');
+    this.logger.verbose('Executing Image test...');
 
     const result = await this.request.get(url);
     const parsedHtml = await this.parseHtml(result);
     const resources = await this.getResources(parsedHtml);
-    const files = this.filterJavaScriptFiles(resources);
-
-    const results = await Promise.all(files.map(async (file) => await this.checkJavaScriptFile(file)));
+    const files = this.filterImageFiles(resources);
+    const results = await Promise.all(files.map(async (file) => await this.checkImageFile(file)));
 
     return new Result(url, results.filter(result => result.resultType !== TYPE_OK).length === 0 ? TYPE_OK : TYPE_INCORRECT, results);
   }
