+++ src/html/JavaScript/Index.js
@@ -1,7 +1,8 @@
 // @flow
-import Result, { TYPE_OK, TYPE_INCORRECT, TYPE_ERROR } from '../pentest/Result';
+import Result, { TYPE_OK, TYPE_INCORRECT, TYPE_ERROR } from '../../pentest/Result';
+import UglifyJS from 'uglify-js';
 
-export default class Image {
+export default class JavaScript {
   request: Object;
   parseHtml: Function;
   getResources: Function;
@@ -21,13 +22,10 @@ export default class Image {
     this.logger = logger;
   }
 
-  filterImageFiles(resources) {
-    const regxJpg = new RegExp('.*\\/([a-z0-9.-])+(\\.jpg)((\\?)?(.*))', 'i');
-    const regxJpeg = new RegExp('.*\\/([a-z0-9.-])+(\\.jpeg)((\\?)?(.*))', 'i');
-    const regxPng = new RegExp('.*\\/([a-z0-9.-])+(\\.png)((\\?)?(.*))', 'i');
-    const regxSvg = new RegExp('.*\\/([a-z0-9.-])+(\\.svg)((\\?)?(.*))', 'i');
+  filterJavaScriptFiles(resources) {
+    const regx = new RegExp('.*\\/([a-z0-9.])+(\\.js)((\\?)?(.*))', 'i');
     return resources.filter((resource) => {
-      return regxJpg.test(resource) || regxJpeg.test(resource) || regxPng.test(resource) || regxSvg.test(resource);
+      return regx.test(resource);
     });
   }
 
@@ -46,7 +44,23 @@ export default class Image {
     return result.response.headers.hasOwnProperty('x-content-type-options');
   }
 
-  async checkImageFile(url: string): Object {
+  hasConsoleLogs(result: Object): boolean {
+    return result.body.indexOf('console.log') !== -1 && result.body.indexOf('console.error') !== -1;
+  }
+
+  isMinified(result: Object): boolean {
+    const r = UglifyJS.minify(result.body, { mangle: false, output: { comments: true } });
+
+    return result.body.length/* - (result.body.length * 0.05)*/ <= r.code.length;
+  }
+
+  isUglified(result: Object): boolean {
+    const r = UglifyJS.minify(result.body);
+  
+    return result.body.length/* - (result.body.length * 0.02)*/ <= r.code.length;
+  }
+
+  async checkJavaScriptFile(url: string): Object {
     const result = await this.request.get(url);
     let tests = [];
 
@@ -58,6 +72,9 @@ export default class Image {
     tests = [...tests, new Result('available', TYPE_OK)];
     tests = [...tests, new Result('cache', this.isCached(result) ? TYPE_OK : TYPE_INCORRECT)];
     tests = [...tests, new Result('x-content-type-options', this.hasXContentTypeOptionsHeader(result) ? TYPE_OK : TYPE_INCORRECT)];
+    tests = [...tests, new Result('console-logs', !this.hasConsoleLogs(result) ? TYPE_OK : TYPE_INCORRECT)];
+    tests = [...tests, new Result('minify', this.isMinified(result) ? TYPE_OK : TYPE_INCORRECT)];
+    tests = [...tests, new Result('uglify', this.isUglified(result) ? TYPE_OK : TYPE_INCORRECT)];
 
     return new Result(url, tests.filter(test => test.resultType !== TYPE_OK).length === 0 ? TYPE_OK : TYPE_INCORRECT, tests);
   }
@@ -68,13 +85,14 @@ export default class Image {
    * @returns {Object}
    */
   async execute(url: string): Object {
-    this.logger.verbose('Executing Image test...');
+    this.logger.verbose('Executing JavaScript test...');
 
     const result = await this.request.get(url);
     const parsedHtml = await this.parseHtml(result);
     const resources = await this.getResources(parsedHtml);
-    const files = this.filterImageFiles(resources);
-    const results = await Promise.all(files.map(async (file) => await this.checkImageFile(file)));
+    const files = this.filterJavaScriptFiles(resources);
+
+    const results = await Promise.all(files.map(async (file) => await this.checkJavaScriptFile(file)));
 
     return new Result(url, results.filter(result => result.resultType !== TYPE_OK).length === 0 ? TYPE_OK : TYPE_INCORRECT, results);
   }
