+++ src/security/FingerPrint/Index.js
@@ -1,5 +1,5 @@
 // @flow
-import Result, { TYPE_OK, TYPE_INCORRECT } from '../pentest/Result';
+import Result, { TYPE_OK, TYPE_INCORRECT } from '../../pentest/Result';
 
 export default class FingerPrint {
   request: Object;
