+++ src/security/XFrameOptions/Index.js
@@ -1,5 +1,5 @@
 // @flow
-import Result, { TYPE_OK, TYPE_ERROR } from '../pentest/Result';
+import Result, { TYPE_OK, TYPE_ERROR } from '../../pentest/Result';
 
 export default class XFrameOptions {
   request: Object;
